backend/requirements.txt
aiosqlite==0.21.0
alembic==1.16.4
asyncpg==0.30.0
bcrypt==4.1.3
celery==5.5.3
cryptography==45.0.7
fastapi==0.115.7
httpx==0.28.1
langgraph==0.2.62
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
pydantic-settings==2.10.1
PyMuPDF==1.26.4
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytest==8.4.1
playwright==1.55.0
playwright-stealth==2.0.0
redis==6.4.0
sqlalchemy==2.0.43
sentry-sdk[fastapi]==2.35.0
structlog==25.4.0
uvicorn[standard]==0.35.0
